once again split down test classes into manageable sizes test_project_views.py ->  test_project_views.py and  test_project_contributor_views.py
